pyproject.toml
[project]
name = "nlhe-refactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dm-tree>=0.1.9",
    "gymnasium>=0.29",
    "jinja2==3.1.4",
    "lz4>=4.4.4",
    "matplotlib>=3.10.6",
    "maturin>=1.9.4",
    "mpmath==1.3.0",
    "networkx==3.3",
    "numpy>=1.24",
    "ray[tune]>=2.49.0",
    "scipy>=1.16.1",
    "sympy==1.13.3",
    "tensorboard>=2.20.0",
    "tqdm>=4.67.1",
    "torch==2.7.1+cu118",
    "torchaudio==2.7.1+cu118",
    "torchvision==0.22.1+cu118",
    "typing-extensions>=4.7 ; python_full_version < '3.11'",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
]


# Pin ONLY these packages to the PyTorch CUDA 11.8 index
[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }

# Declare the extra index as explicit so nothing else falls back to it
[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
